refactor: extract shared rendering helpers and add drift check (M7, M12)
- M7: Extract duplicated helpers (isEmptyValue, isPlaceholderName,
  stripInlineSource, normalizeCell, insufficientDataNotice) into
  dedicated rendering-helpers.ts files. Remove inline copies from
  section-formatter.ts, docx-section-renderers.ts, and researchManager.ts.

- M12: Add scripts/check-shared-drift.sh that compares backend and
  frontend copies of metric-formatter.ts and rendering-helpers.ts,
  failing if they diverge. Can be run in CI to prevent unnoticed drift.
